Login/logout and restricted pages.

+++ studybud/base/views.py
@@ -1,10 +1,43 @@
 # Create your views here.
+from django.contrib import messages
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.models import User
 from django.db.models import Q
 from django.http import HttpRequest, HttpResponse
 from django.shortcuts import render, redirect
 
 from .forms import RoomForm
 from .models import Room, Topic
+
+
+def login_page(request: HttpRequest) -> HttpResponse:
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == 'POST':
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'Username or password is incorrect.')
+
+        user = authenticate(request, username=username, password=password)
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Username or password is incorrect.')
+
+    context = {}
+    return render(request, 'base/login_register.html', context)
+
+
+def logout_user(request: HttpRequest) -> HttpResponse:
+    logout(request)
+    return redirect('home')
 
 
 def home(request: HttpRequest) -> HttpResponse:
@@ -28,6 +61,7 @@ def view_room(request: HttpRequest, pk: str) -> HttpResponse:
     return render(request, 'base/room.html', context)
 
 
+@login_required(login_url='login')
 def create_room(request: HttpRequest) -> HttpResponse:
     form = RoomForm()
     if request.method == 'POST':
@@ -40,9 +74,13 @@ def create_room(request: HttpRequest) -> HttpResponse:
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def update_room(request: HttpRequest, pk: str) -> HttpResponse:
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed to edit this room.')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -54,8 +92,13 @@ def update_room(request: HttpRequest, pk: str) -> HttpResponse:
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def delete_room(request: HttpRequest, pk: str) -> HttpResponse:
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed to edit this room.')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
